Applied an event on action button of MDToolBar. Removed deposit layout in History scene.
