+++ main.py
@@ -1,6 +1,8 @@
+from kivy.properties import ObjectProperty
 from kivy.uix.boxlayout import BoxLayout
 from kivymd.app import MDApp
 from kivy.factory import Factory
+from kivymd.uix.bottomsheet import MDCustomBottomSheet
 from kivymd.uix.list import ILeftBodyTouch
 from kivymd.uix.selectioncontrol import MDCheckbox
 
@@ -10,6 +12,7 @@ from src.common.const import DATABASE_FILENAME
 Factory.register('MainScene', module='src.main_scene')
 Factory.register('HistoryScene', module='src.history_scene')
 Factory.register('SettingScene', module='src.setting_scene')
+Factory.register('DepositSheet', module='src.deposit_sheet')
 
 db = DataBase(DATABASE_FILENAME)
 db.create_new_tables()
@@ -24,13 +27,16 @@ class Root(BoxLayout):
     pass
 
 
+
+
 class MyApp(MDApp):
     dialog = None
+    deposit_sheet = None
 
     def __init__(self, **kwargs):
+        super().__init__(**kwargs)
         self.theme_cls.theme_style = db.get_config()[1]
         self.currency = db.get_config()[2]
-        super().__init__(**kwargs)
 
     def toggle_theme(self, switch, value):
         if value:
@@ -47,7 +53,10 @@ class MyApp(MDApp):
         screen_manager.transition.direction = direction
         screen_manager.current = screen_name
 
+    def show_deposit_sheet(self):
+        self.deposit_sheet = MDCustomBottomSheet(screen=Factory.DepositSheet())
+        self.deposit_sheet.open()
+
 
 if __name__ == '__main__':
     MyApp().run()
-
