+++ my.kv
@@ -2,6 +2,8 @@
 #:include src/kv_files/main_scene.kv
 #:include src/kv_files/history_scene.kv
 #:include src/kv_files/setting_scene.kv
+#:include src/kv_files/deposit_sheet.kv
+
 
 Root:
     orientation: 'vertical'
@@ -13,18 +15,22 @@ Root:
             id: main
         HistoryScene:
             name: "history"
-            id: deposit
+            id: history
         SettingScene:
             name: "setting"
             id: setting
+
     MDBottomAppBar:
         MDToolbar:
             title: screen_manager.current
             icon: "git"
+            on_action_button: app.show_deposit_sheet()
             specific_text_color: [1,1,1,1] if app.theme_cls.theme_style == 'Dark' else [0.2, 0.2, 0.2, 1]
             md_bg_color: app.theme_cls.primary_color
             left_action_items: [['home', lambda x: app.change_screen('main')]]
             right_action_items: [['coins', lambda x: app.change_screen('history', 'right')], ['cogs', lambda x: app.change_screen('setting', 'left')]]
+
+
 
 #            MDIconButton:
 #                icon: 'coins'
